# openapi.yaml
openapi: 3.0.1
info:
  title: gpt-db API
  version: 1.0.0
servers:
  - url: https://gpt-db.vercel.app
    description: Production server
paths:
...
  /food/catalog:
    get:
      summary: List products
      operationId: listProducts
      tags:
        - food
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: false
        - in: query
          name: upc
          schema:
            type: string
          required: false
        - in: query
          name: tag
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Successful response
    post:
      summary: Create or update a product
      operationId: upsertProduct
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Human-readable product name
                upc:
                  type: string
                  description: Barcode used for lookups
                tags:
                  type: array
                  items:
                    type: string
                  description: Category labels for the product
                ingredients:
                  type: array
                  items:
                    type: string
                  description: Ingredient list for the product
                nutrition:
                  type: object
                  description: Per-unit nutrition facts
                  properties:
                    calories:
                      type: number
                      minimum: 0
                      description: Energy in kcal per unit
                    protein:
                      type: number
                      minimum: 0
                      description: Protein (g)
                    fat:
                      type: number
                      minimum: 0
                      description: Fat (g)
                    carbs:
                      type: number
                      minimum: 0
                      description: Carbohydrates (g)
                    fiber:
                      type: number
                      minimum: 0
                      description: Fiber (g)
                    sugars:
                      type: number
                      minimum: 0
                      description: Sugars (g)
                    saturated_fat:
                      type: number
                      minimum: 0
                      description: Saturated fat (g)
                    trans_fat:
                      type: number
                      minimum: 0
                      description: Trans fat (g)
                    cholesterol_mg:
                      type: number
                      minimum: 0
                      description: Cholesterol (mg)
                    sodium_mg:
                      type: number
                      minimum: 0
                      description: Sodium (mg)
                    potassium_mg:
                      type: number
                      minimum: 0
                      description: Potassium (mg)
                    calcium_mg:
                      type: number
                      minimum: 0
                      description: Calcium (mg)
                    iron_mg:
                      type: number
                      minimum: 0
                      description: Iron (mg)
                    magnesium_mg:
                      type: number
                      minimum: 0
                      description: Magnesium (mg)
                    phosphorus_mg:
                      type: number
                      minimum: 0
                      description: Phosphorus (mg)
                    zinc_mg:
                      type: number
                      minimum: 0
                      description: Zinc (mg)
                    selenium_mcg:
                      type: number
                      minimum: 0
                      description: Selenium (mcg)
                    copper_mg:
                      type: number
                      minimum: 0
                      description: Copper (mg)
                    manganese_mg:
                      type: number
                      minimum: 0
                      description: Manganese (mg)
                    vitamin_a_mcg:
                      type: number
                      minimum: 0
                      description: Vitamin A (mcg RAE)
                    vitamin_c_mg:
                      type: number
                      minimum: 0
                      description: Vitamin C (mg)
                    vitamin_d_mcg:
                      type: number
                      minimum: 0
                      description: Vitamin D (mcg)
                    vitamin_e_mg:
                      type: number
                      minimum: 0
                      description: Vitamin E (mg)
                    vitamin_k_mcg:
                      type: number
                      minimum: 0
                      description: Vitamin K (mcg)
                    thiamin_mg:
                      type: number
                      minimum: 0
                      description: Thiamin B1 (mg)
                    riboflavin_mg:
                      type: number
                      minimum: 0
                      description: Riboflavin B2 (mg)
                    niacin_mg:
                      type: number
                      minimum: 0
                      description: Niacin B3 (mg)
                    vitamin_b6_mg:
                      type: number
                      minimum: 0
                      description: Vitamin B6 (mg)
                    folate_mcg:
                      type: number
                      minimum: 0
                      description: Folate (mcg DFE)
                    vitamin_b12_mcg:
                      type: number
                      minimum: 0
                      description: Vitamin B12 (mcg)
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  item:
                    type: object
        '201':
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  item:
                    type: object
  /food/catalog/{product_id}:
    get:
      summary: Retrieve product by ID
      operationId: getProduct
      tags:
        - food
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Product not found
    delete:
      summary: Delete product
      operationId: deleteProduct
      tags:
        - food
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: string
        - in: query
          name: force
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted:
                    type: boolean
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []